+++ index.html
@@ -226,8 +226,8 @@
       </div>
     </section>
   </template>
-
-  <script src="./js/main.js"></script>
+  <script type="module" src="./js/main.js"></script>
+  <!-- <script src="./js/main.js"></script> -->
 
 </body>
 </html>
